Portability fixes as a result of trying to build on SF compile farm. 	Obtained clean builds on all hosts, with some minor tweaking 	required for Apple OS X and SunOS. A serious bug in constructBlock fixed. Placement vector new writes the 	vector size in the first word; it works with gcc 3+ because it 	writes that word before calling the constructors, wheras the 	earlier versions do it after. Well, it is fixed now. Some ctr overloads were conflicting with original algorithms; this includes 	everything that takes a comparator. Use the iterator version for 	clarity. gcc 3+ can figure it out, but there is no reason why it 	should be so taxed. Removed -Weffc++. I KNOW what I do with my pointers, dammit!

+++ memlink.cc
@@ -106,7 +106,7 @@ void memlink::read (istream& is)
     size_t n;
     is >> n;
     assert (n % elementSize() == 0 && "You are trying to read a block with different element type.");
-    size_t btr = min (size_t(n), size());
+    size_t btr = min (n, size());
     is.read (data(), btr);
     resize (btr);
     is.skip (n - btr);
